HP Bar
Added HP Bar to the lower left part of the screen. Also added a heal screen last commit but I forgot to mention it. Looks like the shooter slime is not shooting/dealing damage, will need to understand why. Also the wave manager is part of the UI for some reason. May need to change. - Alex