Grenade
I fixed the huge mess of a script I had and moved the aggro zones into a different layers for the slimes in order to make the grenade not impact with it. Need to make a new explosion prefab because the other one's range is too huge + check random impacts.